% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIncidence.r
\name{ageadjust}
\alias{ageadjust}
\title{sum age adjusted incidence proportion and calculate 95% lowerCI and 95% upperCI}
\usage{
ageadjust(incidencePropdata, alpha = 0.05)
}
\arguments{
\item{incidencePropdata}{output of Argos packages calculateIncidence code if standardization == 'direct'}

\item{alpha}{0.05}
}
\description{
sum age adjusted incidence proportion and calculate 95% lowerCI and 95% upperCI
}
