% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIncidence.r
\name{agespe}
\alias{agespe}
\title{aggregate incidence proportion according to gender and age section}
\usage{
agespe(incidencePropdata)
}
\arguments{
\item{incidencePropdata}{output of Argos packages calculateIncidence code}
}
\description{
aggregate incidence proportion according to gender and age section
}
