\name{Argos-package}
\alias{Argos-package}
\alias{Argos}
\docType{package}
\title{
\packageTitle{Argos}
}
\description{
\packageDescription{Argos}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Argos}
\packageIndices{Argos}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Argos}

Maintainer: \packageMaintainer{Argos}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
