% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos_translate.R
\name{argos_translate_only}
\alias{argos_translate_only}
\alias{argos_translate}
\title{Translate via Argos}
\usage{
argos_translate_only(text, code_from = "en", code_to = "es")

argos_translate(
  text,
  code_from = "en",
  code_to = "es",
  update_package_index = TRUE
)
}
\arguments{
\item{text}{the text to translate in language \code{code_from} into \code{code_to}}

\item{code_from}{Code for the language text is coming from}

\item{code_to}{Code for the language text is going to}

\item{update_package_index}{Should package index be updated prior to
running install_language}
}
\value{
The text from the translation
}
\description{
Translate via Argos
}
\note{
This only translates the text, it does not install language packages.
}
\examples{
if (is_argos_installed()) {
  argos_translate("Hello World")
  argos_translate_only("Hello World")
}
}
