% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIncidence.r
\name{bybirth}
\alias{bybirth}
\title{aggregate incidence proportion according to gender and birth year section}
\usage{
bybirth(incidencePropdata)
}
\arguments{
\item{incidencePropdata}{output of Argos packages calculateIncidence code}
}
\description{
aggregate incidence proportion according to gender and birth year section
}
