% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_monthly_incidence.R
\name{cal_incidencerate}
\alias{cal_incidencerate}
\title{calculate incidence rate}
\usage{
cal_incidencerate(
  startDate = as.Date("2003-01-01"),
  endDate = as.Date("2013-12-31"),
  timeInterval = "months",
  final_pop = final_pop
)
}
\arguments{
\item{endDate}{}

\item{timeInterval}{minumal unit for cohort start date ('year' > 'quarter' > 'month' > 'week' > day')}

\item{final_pop}{}
}
\description{
calculate incidence rate
}
