% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{calculateNumberPerCovTime}
\alias{calculateNumberPerCovTime}
\title{function to calculate number per covariate per time unit}
\usage{
calculateNumberPerCovTime(plpData, population = NULL, minDateUnit = "year")
}
\arguments{
\item{cohort}{cohort should be data frame with at least four columns 'rowId', 'subjectId', 'cohortId', 'cohortStartDate'}
}
\description{
function to calculate number per covariate per time unit
}
