% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSurvival.r
\name{calculateSurvival}
\alias{calculateSurvival}
\title{Calculating survival rate function}
\usage{
calculateSurvival(survivalData = survivalData, Agedivided = TRUE,
  AgeSet = list(30:39, 40:49, 50:59, 60:69, 70:79, 80:99),
  genderSet = list(8507, 8532), startYearSet = startYearSet,
  observationEndYear = 2013)
}
\arguments{
\item{survivalData}{outcome of Argos package readySurvData code}

\item{Agedivided}{TRUE then total patients survival rate will calculated, FALSE then age specified survival rate will calculated}

\item{AgeSet}{30-39, 40-49, 50-59, 60-69, 70-79, 80-99 will set as default}

\item{genderSet}{8507 as male, 8532 as female}

\item{startYearSet}{}
}
\description{
Calculating survival rate function
}
