% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackage.r
\name{check.packages}
\alias{check.packages}
\title{check and ready packages needed}
\usage{
check.packages(pkgname)
}
\arguments{
\item{pkgname}{package name}
}
\description{
check and ready packages needed
}
