% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_argos_condaenv.R
\name{create_argos_condaenv}
\alias{create_argos_condaenv}
\title{Install Argos Conda Environment}
\usage{
create_argos_condaenv(envname = "argos", packages = "argostranslate", ...)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{packages}{character vector of packages to install.  ``argostranslate"`
is automatically added to the vector, passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{...}{additional arguments to pass to  \code{\link[reticulate:conda-tools]{reticulate::conda_create()}}.}
}
\value{
Returns the path to the Python binary associated with the
newly-created conda environment.
}
\description{
Install Argos Conda Environment
}
\examples{
\dontrun{
  if (!is_argos_installed()) {
    create_argos_condaenv()
  }
}
}
