% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.r
\name{createStandSettings}
\alias{createStandSettings}
\title{Creating setting for age/gender/race/ethnicity standardization}
\usage{
createStandSettings(ageStandardization = TRUE,
  ageGroupStandardization = FALSE, genderStandardization = TRUE,
  raceStandardization = FALSE, ethnicityStandardization = FALSE,
  mode = "direct", basePopulation, year = 2007)
}
\description{
Creating setting for age/gender/race/ethnicity standardization
}
