% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createYearMonthTable.R
\name{createYearMonthTable}
\alias{createYearMonthTable}
\title{Create a table of database observation start and end date.}
\usage{
createYearMonthTable(
  startDate = lubridate::ymd("20180101"),
  endDate = lubridate::ymd("20220401")
)
}
\arguments{
\item{startDate}{the database's observation start date}

\item{endDate}{the database's observation end date}
}
\description{
Create a table of database observation start and end date.
}
