% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createYearMonthTable.R
\name{days_in_month_argos}
\alias{days_in_month_argos}
\title{calculating days in the month}
\usage{
days_in_month_argos(date)
}
\arguments{
\item{date}{}
}
\description{
calculating days in the month
}
