% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costCalculation.r
\name{extractVisitCost}
\alias{extractVisitCost}
\title{get cost data}
\usage{
extractVisitCost(connectionDetails, cdmDatabaseSchema,
  cohortDatabaseSchema, cohortTable, cohortId = -1,
  costWindowStart = -60, costWindowEnd = 365,
  minCostDateUnit = "quarter")
}
\arguments{
\item{cohortId}{target Cohort Id, if this is -1, it will extract cost data for whole subject in the cohortTable}

\item{minCostDateUnit}{minumal unit for cohort start date ('year' > 'quarter' > 'month' > 'week' > day')}
}
\description{
get cost data
}
