% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsurvData.r
\name{getsurvData}
\alias{getsurvData}
\title{get outcome, survivaltime, needed variables}
\usage{
getsurvData(plpData = plpData, population = population)
}
\arguments{
\item{plpData}{outcome of PatientLevelPrediction package getPlpData code}

\item{population}{outcome of PatientLevelPrediction package createStudyPopulation code}
}
\description{
get outcome, survivaltime, needed variables
}
