% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_argos.R
\name{install_argos}
\alias{install_argos}
\alias{is_argos_installed}
\title{Install Argos}
\usage{
install_argos(packages = NULL, pip = TRUE, ...)

is_argos_installed()
}
\arguments{
\item{packages}{character vector of packages to install.  ``argostranslate"`
is automatically added to the vector, passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{pip}{Boolean; use pip for package installation?
Passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{...}{additional arguments to pass to  \code{\link[reticulate:py_install]{reticulate::py_install()}}.}
}
\value{
Nothing (\code{invisible(NULL)}).  Command is called for side effects
}
\description{
Install Argos
}
\examples{
\dontrun{
  if (!is_argos_installed()) {
    install_argos()
  }
}
}
