% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{limitCovariatesToPopulation}
\alias{limitCovariatesToPopulation}
\title{function to limit covariates of plpData to population}
\usage{
limitCovariatesToPopulation(covariates, rowIds)
}
\description{
function to limit covariates of plpData to population
}
