% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_language_package.R
\name{list_installed_language_packages}
\alias{list_installed_language_packages}
\title{List Installed Argos Language Packages}
\usage{
list_installed_language_packages()
}
\value{
If none installed, \code{NULL} and if some are installed, a \code{data.frame}
of information of the language packags
}
\description{
List Installed Argos Language Packages
}
\examples{
if (is_argos_installed()) {
  list_installed_language_packages()
}
}
