% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{loadDisabilityWeight}
\alias{loadDisabilityWeight}
\title{Load default disability weight  data}
\usage{
loadDisabilityWeight(country, year)
}
\arguments{
\item{country}{'IHME' or or 3-digit Coutry code (eg 'KOR' = 'Republic of Korea')}
}
\description{
Load default disability weight  data
}
