% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{loadLifeExpectancy}
\alias{loadLifeExpectancy}
\title{Load default life expectancy data}
\usage{
loadLifeExpectancy(country)
}
\arguments{
\item{country}{'WHO' for world standard life expectancy or 3-digit Coutry code (eg 'KOR' = 'Republic of Korea')}
}
\description{
Load default life expectancy data
}
