% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{loadSurvivalExpectancy}
\alias{loadSurvivalExpectancy}
\title{Load default mid-year population data}
\usage{
loadSurvivalExpectancy(country)
}
\arguments{
\item{country}{'WHO' for world standard population publisehd by WHO, or 3-digit Coutry code (eg 'KOR' = 'Republic of Korea')}
}
\description{
Load default mid-year population data
}
