% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforMortality.r
\name{PlotByBirthMort}
\alias{PlotByBirthMort}
\title{plot mortality proportion}
\usage{
PlotByBirthMort(mortalityPropdata, title, outputFolder, fileName,
  imageExtension = "png")
}
\description{
plot mortality proportion
}
