% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforMortality.r
\name{PlotByDiagnosisMort}
\alias{PlotByDiagnosisMort}
\title{plot mortality proportion}
\usage{
PlotByDiagnosisMort(incidencePropdata, ageSpetitle, ageAdjtitle, outputFolder,
  ageSpefileName, ageAdjfileName, imageExtension = "png")
}
\description{
plot mortality proportion
}
