% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforCost.r
\name{plotforCostPerMt}
\alias{plotforCostPerMt}
\title{monthly cost from 2months before diagnosis to 12months after diagnosis}
\usage{
plotforCostPerMt(costData)
}
\arguments{
\item{costData}{outcome of Argos package extractVisitCost code if minCostDateUnit = "month"}
}
\description{
monthly cost from 2months before diagnosis to 12months after diagnosis
}
