% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforCost.r
\name{plotforCostPerYrBarPat}
\alias{plotforCostPerYrBarPat}
\title{the cost of the diagnosed year paid by patient}
\usage{
plotforCostPerYrBarPat(costData)
}
\description{
the cost of the diagnosed year paid by patient
}
