% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforCost.r
\name{plotforCostPerYrdiv}
\alias{plotforCostPerYrdiv}
\title{the total cost of the diagnosed year}
\usage{
plotforCostPerYrdiv(costData)
}
\arguments{
\item{costData}{outcome of Argos package extractVisitCost code if minCostDateUnit = "year"}
}
\description{
the total cost of the diagnosed year
}
