% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforDALY.r
\name{plotforDALY}
\alias{plotforDALY}
\title{DALY value plot according to diagnosis year}
\usage{
plotforDALY(DALYdata, diseaseList = diseaseList)
}
\arguments{
\item{DALYdata}{outcome of Argos package DALY code}
}
\description{
DALY value plot according to diagnosis year
}
