% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforDALY.r
\name{plotforDALYratio}
\alias{plotforDALYratio}
\title{YLL:YLD ratio according to diagnosis year}
\usage{
plotforDALYratio(DALYdata, diseaseList = diseaseList)
}
\arguments{
\item{DALYdata}{outcome of Argos package DALY code}
}
\description{
YLL:YLD ratio according to diagnosis year
}
