% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforSurvival.r
\name{plotSurvival1Yr}
\alias{plotSurvival1Yr}
\title{1year survival rate plot by diagnosis year}
\usage{
plotSurvival1Yr(agedivSurvCal, diseaseList = diseaseList)
}
\arguments{
\item{agedivSurvCal}{outcome of Argos package calculateSurvival code if Agediv = TRUE}
}
\description{
1year survival rate plot by diagnosis year
}
