% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforSurvival.r
\name{plotSurvival3Yr}
\alias{plotSurvival3Yr}
\title{3year survival rate plot by diagnosis year}
\usage{
plotSurvival3Yr(agedivSurvCal)
}
\arguments{
\item{agedivSurvCal}{outcome of Argos package calculateSurvival code if Agediv = TRUE}
}
\description{
3year survival rate plot by diagnosis year
}
