% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforSurvival.r
\name{plotSurvival5Yr}
\alias{plotSurvival5Yr}
\title{5year survival rate plot by diagnosis year}
\usage{
plotSurvival5Yr(agedivSurvCal)
}
\arguments{
\item{agedivSurvCal}{outcome of Argos package calculateSurvival code if Agediv = TRUE}
}
\description{
5year survival rate plot by diagnosis year
}
