% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforSurvival.r
\name{plotSurvivalTotal}
\alias{plotSurvivalTotal}
\title{total patients's survival rate plot by diagnosis year}
\usage{
plotSurvivalTotal(totalSurvCal)
}
\arguments{
\item{totalSurvCal}{outcome of Argos package calculateSurvival code if Agediv = FALSE}
}
\description{
total patients's survival rate plot by diagnosis year
}
