% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSurvival.r
\name{readySurvData}
\alias{readySurvData}
\title{get outcome  data}
\usage{
readySurvData(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  outcomeDatabaseSchema, cohortTable, covariateSettings, targetCohortId,
  outcomeId, requireTimeAtRisk = FALSE, minTimeAtRisk = 1,
  riskWindowStart = 0, riskWindowEnd = 365 * 5,
  removeSubjectsWithPriorOutcome = TRUE, minDateUnit = "year")
}
\arguments{
\item{targetCohortId}{cancer type cohort id}

\item{outcomeId}{outcome = death}

\item{requireTimeAtRisk}{must be FALSE (default)}

\item{riskWindowEnd}{if you want to see max 5-year survival rate then 365*5}

\item{removeSubjectsWithPriorOutcome}{}

\item{minDateUnit}{minumal unit for cohort start date ('year' > 'quarter' > 'month' > 'day')}
}
\description{
get outcome  data
}
