% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLoadArgos.R
\name{saveArgosData}
\alias{saveArgosData}
\title{Save the cohort data to folder}
\usage{
saveArgosData(argosData, file, envir = NULL, overwrite = F)
}
\arguments{
\item{argosData}{An object of type \code{argosData} as generated using
\code{getArgosData}.}

\item{file}{The name of the folder where the data will be written. The folder should
not yet exist.}

\item{envir}{The environment for to evaluate variables when saving}

\item{overwrite}{Whether to force overwrite an existing file}
}
\description{
\code{saveArgosData} saves an object of type argosData to folder.
}
\details{
The data will be written to a set of files in the folder specified by the user.
}
\examples{
# todo

}
