% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCost.r
\name{savecost}
\alias{savecost}
\title{saving tables and plot image which are the results cost analysis}
\usage{
savecost(outputFolder, PlottotalcostperYrdiv, plotperYrBarplotPayer,
  plotperYrBarplotPatient, imageExtension = "png")
}
\arguments{
\item{PlottotalcostperYrdiv}{}

\item{plotperYrBarplotPayer}{}

\item{plotperYrBarplotPatient}{}

\item{imageExtension}{plot file's extension}
}
\description{
saving tables and plot image which are the results cost analysis
}
