% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDALY.r
\name{saveDALY}
\alias{saveDALY}
\title{saving plot images which are the results DALY calculate}
\usage{
saveDALY(outputFolder, plotDALY, plotDALYratio, imageExtension = "png")
}
\arguments{
\item{imageExtension}{plot file's extension}

\item{plotforDALY}{}

\item{plotforDALYratio}{}
}
\description{
saving plot images which are the results DALY calculate
}
