% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveIncidence.r
\name{saveIncidence}
\alias{saveIncidence}
\title{saving tables and plot image which are the results incidence analysis}
\usage{
saveIncidence()
}
\description{
saving tables and plot image which are the results incidence analysis
}
