% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSurvival.r
\name{saveSurvival}
\alias{saveSurvival}
\title{saving tables and plot image which are the results incidence analysis}
\usage{
saveSurvival(outputFolder, plot1yrsurvival, plot3yrsurvival,
  plot5yrsurvival, plottotalsurvival, imageExtension = "png")
}
\arguments{
\item{outputFolder}{}

\item{imageExtension}{plot file's extension}
}
\description{
saving tables and plot image which are the results incidence analysis
}
