% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsurvData.r
\name{survivalCal}
\alias{survivalCal}
\title{calculate survival rate using survival package}
\usage{
survivalCal(survivalDuration = survivalTime,
  outcomeCount = outcomeCount,
  survivalDurationTime = survivalDurationTime)
}
\arguments{
\item{survivalDuration}{time gap between first diagnosis date and death date or last observation date (survivalTime)}

\item{outcomeCount}{binary value (if outcome = death, death = 1 and alive = 0)}

\item{survivalDurationTime}{n-year survival -> 365*n}
}
\description{
calculate survival rate using survival package
}
