% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableIncidence.R
\name{tableAgeAdjusted}
\alias{tableAgeAdjusted}
\title{make age adjusted incidence rate table}
\usage{
tableAgeAdjusted(ageadjIncData)
}
\arguments{
\item{ageadjIncData}{result of ageadjust code}
}
\description{
make age adjusted incidence rate table
}
