% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableIncidence.R
\name{tableAgeSpecified}
\alias{tableAgeSpecified}
\title{make age specified incidence rate table}
\usage{
tableAgeSpecified(ageSpecifiedData)
}
\arguments{
\item{ageSpecifiedData}{result of agespe code}
}
\description{
make age specified incidence rate table
}
