% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableIncidence.R
\name{tablebirthCohortInc}
\alias{tablebirthCohortInc}
\title{make incidence rate by birth cohort table}
\usage{
tablebirthCohortInc(birthIncData)
}
\arguments{
\item{birthIncData}{result of bybirth code}
}
\description{
make incidence rate by birth cohort table
}
