% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSurvival.r
\name{tableSurvTotal}
\alias{tableSurvTotal}
\title{make total survival rate table}
\usage{
tableSurvTotal(totalSurvData)
}
\arguments{
\item{totalSurvData}{result of calculateSurvival code Agedivided = FALSE}
}
\description{
make total survival rate table
}
