% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAPC.R
\name{ageAdjAPC}
\alias{ageAdjAPC}
\title{calculate Annual percent change of age adjusted incidence proportion}
\usage{
ageAdjAPC(ageAdjustedInc = ageadjInc)
}
\arguments{
\item{ageAdjustedInc}{}
}
\description{
calculate Annual percent change of age adjusted incidence proportion
}
