% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DALY.r
\name{calculateDALY}
\alias{calculateDALY}
\title{Calculate Years of Life Lost (YLL), Years Lost due to Disability (YDD), and Disability-Adjusted Life Year (DALY)}
\usage{
calculateDALY(outcomeData, refLifeExpectancy, disabilityWeight = 0.5,
  outcomeDisabilityWeight = 1, minTimeAtRisk = 1,
  observeStartYr = NULL, observeEndYr = NULL, discount = 0.3,
  ageWeighting = TRUE, outputFolder)
}
\arguments{
\item{outcomeData}{outcomeData generated by Argos package.}

\item{refLifeExpectancy}{reference  life expectancy data to calculate YDD. The default life expectancy table is provided and can be loaded via loadLifeExpectancy function}

\item{disabilityWeight}{disability weight for the certain condition of plpData to calculate YLD. if this is 0, YLD is 0. It should be float value (0~1).}

\item{outcomeDisabilityWeight}{Disability weight for outcome.  If the outcome is death, then disability weight should be 1. It should be float value (0~1).}

\item{minTimeAtRisk}{time at risk (days). usually it should be identical to the minTAR of plpData}

\item{observeStartYr}{From which year, the observation of DALY should be started?}

\item{observeEndYr}{Until which year, the observation of DALY should be ended?}

\item{discount}{discount value, float(0~1). default value is 0.3.}

\item{outputFolder}{outputFolder}

\item{ageWeghting}{logical value (TRUE or FALSE).}
}
\description{
Calculate Years of Life Lost (YLL), Years Lost due to Disability (YDD), and Disability-Adjusted Life Year (DALY)
}
