% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DALY.r
\name{calculateDALY}
\alias{calculateDALY}
\title{Calculate Years of Life Lost (YLL), Years Lost due to Disability (YDD), and Disability-Adjusted Life Year (DALY)}
\usage{
calculateDALY(plpData, population, refLifeExpectancy,
  disabilityWeight = 0, minTimeAtRisk, discount, outputFolder)
}
\arguments{
\item{plpData}{plpData generated by PLP package.}

\item{population}{population data generated by PLP package. the outcome should be death}

\item{refLifeExpectancy}{reference  life expectancy data to calculate YDD}

\item{disabilityWeight}{disability weight for the certain condition of plpData to calculate YLD. if this is 0, YLD is 0}

\item{minTimeAtRisk}{time at risk (days). usually it should be identical to the minTAR of plpData}

\item{discount}{discount count, float(0~1). default value is 0.3}

\item{outputFolder}{outputFolder}
}
\description{
Calculate Years of Life Lost (YLL), Years Lost due to Disability (YDD), and Disability-Adjusted Life Year (DALY)
}
