% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSurvival.r
\name{calculateSurvival}
\alias{calculateSurvival}
\title{Calculating survival rate function}
\usage{
calculateSurvival(survivalData = survivalData, refPopulation = refPop,
  Agedivided = Agedivided, AgeSet = list(30:39, 40:49, 50:59, 60:69,
  70:79, 80:99), genderSet = list(8507, 8532),
  startYearSet = list(2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009,
  2010, 2011, 2012), birthYearSet = list(1960:1964, 1965:1969, 1970:1974,
  1975:1979, 1980:1984, 1985:1989), observationEndYear)
}
\arguments{
\item{survivalData}{outcome of Argos package readySurvData code}

\item{Agedivided}{TRUE then total patients survival rate will calculated, FALSE then age specified survival rate will calculated}

\item{AgeSet}{30-39, 40-49, 50-59, 60-69, 70-79, 80-99 will set as default}

\item{genderSet}{8507 as male, 8532 as female}

\item{startYearSet}{}
}
\description{
Calculating survival rate function
}
