% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIncidence.r
\name{getIncidenceData}
\alias{getIncidenceData}
\title{get incidence data}
\usage{
getIncidenceData(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  outcomeDatabaseSchema, cohortTable, targetCohortId, minDateUnit = "year")
}
\arguments{
\item{cdmDatabaseSchema}{}

\item{minDateUnit}{minumal unit for cohort start date ('year' > 'quarter' > 'month' > 'day')}
}
\description{
get incidence data
}
