% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforIncidence.r
\name{PlotByBirthInc}
\alias{PlotByBirthInc}
\title{incidence proportion plot by birth Year}
\usage{
PlotByBirthInc(birthcohortIncData)
}
\arguments{
\item{birthcohortIncData}{output of Argos packages bybirth code}
}
\description{
incidence proportion plot by birth Year
}
