% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforIncidence.r
\name{PlotByDiagnosisIncAgeAd}
\alias{PlotByDiagnosisIncAgeAd}
\title{Age adjusted incidence proportion plot by diagnosis year}
\usage{
PlotByDiagnosisIncAgeAd(ageadjustIncData)
}
\arguments{
\item{ageadjustIncData}{output of Argos packages ageadjust code}
}
\description{
Age adjusted incidence proportion plot by diagnosis year
}
