% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforIncidence.r
\name{PlotByDiagnosisIncAgeAd}
\alias{PlotByDiagnosisIncAgeAd}
\title{Age adjusted incidence proportion plot by diagnosis year}
\usage{
PlotByDiagnosisIncAgeAd(incidencePropdata)
}
\arguments{
\item{incidencePropdata}{}
}
\description{
Age adjusted incidence proportion plot by diagnosis year
}
