% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforIncidence.r
\name{PlotByDiagnosisIncAgeS}
\alias{PlotByDiagnosisIncAgeS}
\title{Age specified incidence proportion plot by diagnosis year}
\usage{
PlotByDiagnosisIncAgeS(incidencePropdata)
}
\arguments{
\item{incidencePropdata}{outcome of Argos package getIncidenceData code}
}
\description{
Age specified incidence proportion plot by diagnosis year
}
