% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforIncidence.r
\name{PlotByDiagnosisIncAgeS}
\alias{PlotByDiagnosisIncAgeS}
\title{Age specified incidence proportion plot by diagnosis year}
\usage{
PlotByDiagnosisIncAgeS(agespecifiedIncData, diseaseList = diseaseList)
}
\arguments{
\item{agespecifiedIncData}{output of Argos packages agespe code}

\item{diseaseList}{cancerList or psychoticList}
}
\description{
Age specified incidence proportion plot by diagnosis year
}
