% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforCost.r
\name{plotforCostPerMt}
\alias{plotforCostPerMt}
\title{monthly cost from 2months before diagnosis to 12months after diagnosis}
\usage{
plotforCostPerMt(costData)
}
\description{
monthly cost from 2months before diagnosis to 12months after diagnosis
}
