% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforCost.r
\name{plotforCostPerYrBarPat}
\alias{plotforCostPerYrBarPat}
\title{the cost of the diagnosed year paid by patient}
\usage{
plotforCostPerYrBarPat(costData, diseaseList = diseaseList)
}
\arguments{
\item{costData}{outcome of Argos package extractVisitCost code if minCostDateUnit = "year"}
}
\description{
the cost of the diagnosed year paid by patient
}
