% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotforCost.r
\name{plotforCostPerYrBarPay}
\alias{plotforCostPerYrBarPay}
\title{the cost of the diagnosed year paid by payer}
\usage{
plotforCostPerYrBarPay(costData)
}
\arguments{
\item{costData}{outcome of Argos package extractVisitCost code if minCostDateUnit = "year"}
}
\description{
the cost of the diagnosed year paid by payer
}
