% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSurvival.r
\name{saveSurvival}
\alias{saveSurvival}
\title{saving tables and plot image which are the results incidence analysis}
\usage{
saveSurvival()
}
\description{
saving tables and plot image which are the results incidence analysis
}
